+++ src/pages/Executors.vue
@@ -10,11 +10,30 @@
     </q-breadcrumbs>
 
     <q-table
-      title="Executors"
       :data="allExecutorsList"
       :columns="columns"
       row-key="name"
+      :rows-per-page-options="rowsPerPage"
+      :pagination.sync="pagination"
+      :filter="filter"
     >
+      <div slot="top-left" slot-scope="props" class="row items-center">
+        <div class="col-auto q-headline">
+          Executors
+        </div>
+
+        <div class="q-ml-md row">
+          <q-search
+            v-if="allExecutorsList.length"
+            class="col"
+            color="secondary"
+            v-model="filter"
+            placeholder="Filter"
+            clearable
+          />
+        </div>
+      </div>
+
       <div slot="top-right" slot-scope="props" class="fit">
         <executor-new-button v-on:created="created"/>
       </div>
@@ -67,7 +86,12 @@
             align: 'left',
             sortable: true
           }
-        ]
+        ],
+        rowsPerPage: [0],
+        pagination: {
+          rowsPerPage: 0
+        },
+        filter: '',
       }
     },