+++ src/pages/MissionDef.vue
@@ -61,11 +61,30 @@
       </q-card>
 
       <q-table
-        title="Parameters"
         :data="myParameters"
         :columns="columns"
         row-key="name"
+        :rows-per-page-options="rowsPerPage"
+        :pagination.sync="pagination"
+        :filter="filter"
       >
+        <div slot="top-left" slot-scope="props" class="row items-center">
+          <div class="col-auto q-headline">
+            Parameters
+          </div>
+
+          <div class="q-ml-md row">
+            <q-search
+              v-if="myParameters.length"
+              class="col"
+              color="secondary"
+              v-model="filter"
+              placeholder="Filter"
+              clearable
+            />
+          </div>
+        </div>
+
         <div slot="top-right" slot-scope="props" class="fit">
           <parameter-new-button
             :executor-id="params.executorId"
@@ -144,6 +163,11 @@
             sortable: true
           }
         ],
+        rowsPerPage: [0],
+        pagination: {
+          rowsPerPage: 0
+        },
+        filter: '',
       }
     },